[fix] Gather parallel-scan ROWNUM code into px_scan_instnum.{hpp,cpp}
The instnum support was scattered: a dead atomic_instnum class in px_interrupt.hpp, an inline
helper parked in the checker header only to cross the cs/sa-vs-server link boundary, a ~130 line
eligibility block inside the 1181-line checker, and the numbering runtime spread through the
result handler.

Put all of it in one pair. The two halves are split by build mode rather than by file, the way
this directory already does it: eligibility is compiled for the client, and the numbering runtime
- which reaches into the query engine (qfile_*, fetch_peek_dbval) the client does not link - is
guarded to SERVER_MODE/SA_MODE. libcubridcs.so gains no server symbols.

px_scan_checker.cpp loses 130 lines, px_scan_result_handler.cpp loses 102, and the server stops
including a client-only header.
